+++ khiops/core/dictionary.py
@@ -186,7 +186,7 @@ class DictionaryDomain(KhiopsJSONObject):
 
         Parameters
         ----------
-        dictionary : `DictionaryDomain`
+        dictionary : `Dictionary`
             The dictionary to be added.
 
         Raises