+++ khiops/core/helpers.py
@@ -8,6 +8,7 @@
 import os
 import platform
 import subprocess
+import warnings
 
 import khiops.core.internals.filesystems as fs
 from khiops.core import api
@@ -19,6 +20,7 @@ from khiops.core.dictionary import (
 )
 from khiops.core.internals.common import (
     create_unambiguous_khiops_path,
+    deprecation_message,
     is_list_like,
     type_error_message,
 )
@@ -29,6 +31,22 @@ def build_multi_table_dictionary_domain(
 ):
     """Builds a multi-table dictionary domain from a dictionary with a key
 
+    .. note::
+
+        This is a special-purpose function whose goal is to assist in preparing the
+        coclustering deployment.
+
+        This function builds a new root dictionary and adds it to an existing dictionary
+        domain.
+        The new root dictionary only contains one field, which references a preexisting
+        dictionary from the input dictionary domain as a new (secondary) Table variable.
+        The preexisting dictionary must have a key set on it, as this is the join key
+        with the new root table.
+
+    .. warning::
+        This method is *deprecated* since Khiops 10.3.1.0 and will be removed in Khiops
+        11.
+
     Parameters
     ----------
     dictionary_domain : `.DictionaryDomain`
@@ -37,6 +55,11 @@ def build_multi_table_dictionary_domain(
         Name for the new root dictionary
     secondary_table_variable_name : str
         Name, in the root dictionary, for the "table" variable of the secondary table.
+
+    Returns
+    -------
+    `.DictionaryDomain`
+        The new dictionary domain
 
     Raises
     ------
@@ -47,6 +70,16 @@ def build_multi_table_dictionary_domain(
         - the dictionary domain doesn't contain at least a dictionary
         - the dictionary domain's root dictionary doesn't have a key set
     """
+    # Warn the user that this helper function is deprecated and will be removed
+    warnings.warn(deprecation_message("build_multi_table_dictionary_domain", "11.0.0"))
+    return _build_multi_table_dictionary_domain(
+        dictionary_domain, root_dictionary_name, secondary_table_variable_name
+    )
+
+
+def _build_multi_table_dictionary_domain(
+    dictionary_domain, root_dictionary_name, secondary_table_variable_name
+):
     # Check that `dictionary_domain` is a `DictionaryDomain`
     if not isinstance(dictionary_domain, DictionaryDomain):
         raise TypeError(
@@ -266,7 +299,7 @@ def deploy_coclustering(
     # Create a root dictionary containing the keys
     root_dictionary_name = "CC_" + dictionary_name
     table_variable_name = "Table_" + dictionary_name
-    domain = build_multi_table_dictionary_domain(
+    domain = _build_multi_table_dictionary_domain(
         tmp_domain, root_dictionary_name, table_variable_name
     )